updating feat org for easier nav
