+++ src/finagg/sec/feat/refined/__init__.py
@@ -1,0 +1,4 @@
+"""Refined SEC features (features aggregated from raw tables)."""
+
+from .annual_ import Annual
+from .quarterly_ import Quarterly
